requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
python-dotenv==0.19.0
requests==2.31.0
SQLAlchemy==2.0.21
pandas==2.1.0
scikit-learn==1.3.0
beautifulsoup4==4.12.2
nltk==3.8.1
redis==5.0.0
aiohttp==3.8.5
gunicorn==21.2.0
psycopg2-binary==2.9.7
